+++ duckscript_sdk/src/sdk/std/fs/mod.rs
@@ -1,3 +1,5 @@
+mod basename;
+mod canonical;
 mod mkdir;
 mod print;
 mod read;
@@ -12,6 +14,8 @@ static PACKAGE: &str = "fs";
 pub(crate) fn load(commands: &mut Commands, parent: &str) -> Result<(), ScriptError> {
     let package = pckg::concat(parent, PACKAGE);
 
+    commands.set(basename::create(&package))?;
+    commands.set(canonical::create(&package))?;
     commands.set(mkdir::create(&package))?;
     commands.set(print::create(&package))?;
     commands.set(read::create(&package))?;